Use BitSet and move decode to a test utils class
