+++ core/src/main/java/com/microsoft/applicationinsights/internal/statsbeat/NetworkStatsbeat.java
@@ -95,7 +95,6 @@ public class NetworkStatsbeat extends BaseStatsbeat {
         }
     }
 
-    // TODO create a new instrumentation set automatically when its capacity reaches 64
     public void addInstrumentation(String instrumentation) {
         synchronized (lock) {
             current.instrumentationList.add(instrumentation);
